INT-2319 polishing with PR comments
